fs: littlefs: Add Kconfig option to set LittleFS init priority
The default value of 85 was picked to be higher than
SPI_NOR_INIT_PRIORITY=80 so that auto mount option does not fail because
a NOR SPI flash is not yet accessible. It was also picked to be lower
than NET_INIT_PRIO=90 so that interfaces relying on a filesystem (e.g.
through the settings subsystem) can properly initialize.